User Card Assignment
Learned how to use Position:absolute and relative. I did not know that position relative should be attached to the parent div and you will be able to manipulate the child element with position absolute. Given you would want to move the child around the parent element, which is exactly what i needed for this assingment.